test: move get_changed_items tests to test_git_diff_utils.py and add validate_git_compare_ref tests
